# common/api/marly-api-spec.yml
openapi: 3.0.0
info:
  title: Pipeline API
  description: API for running and retrieving results from data processing pipelines
  version: 1.0.0

paths:
  /pipelines:
    post:
      summary: Run pipeline
      operationId: runPipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineRequestModel"
      responses:
        "200":
          description: Pipeline processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  message:
                    type: string
        "500":
          $ref: "#/components/responses/InternalServerError"

  /pipelines/{task_id}:
    get:
      summary: Get pipeline results
      operationId: getPipelineResults
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pipeline results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResult"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    PipelineRequestModel:
      type: object
      properties:
        workloads:
          type: array
          items:
            type: object
            properties:
              pdf_stream:
                type: string
                format: binary
              schemas:
                type: array
                items:
                  type: string
        provider_type:
          type: string
        provider_model_name:
          type: string
        api_key:
          type: string
        additional_params:
          type: object
          additionalProperties: true

    PipelineResult:
      type: object
      properties:
        task_id:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        results:
          type: array
          items:
            $ref: "#/components/schemas/SchemaResult"
        total_run_time:
          type: string

    SchemaResult:
      type: object
      properties:
        schema_id:
          type: string
        metrics:
          type: object
          additionalProperties:
            type: string
        schema_data:
          type: object
          additionalProperties:
            type: string

    JobStatus:
      type: string
      enum:
        - PENDING
        - IN_PROGRESS
        - COMPLETED
        - FAILED

  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string